requirements.txt
faster-whisper
sentence-transformers
faiss-cpu
transformers
torch
tqdm
